Add files for npc
In the previous commit, the appearance of npc use the appearance of weaponsellernpc. In this commit, the appearance for pipedealnpc only was created, and the PipeDealerNPC has been added into the deniran.xml, which means he is visible in the game now.

Change File: PipeDealerNPC.java, pipedealernpc.png, deniran.xml
Issue: #11